Add new interfaces: LikesDao, GenreDao, MpaDao, FriendDao.
