+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -28,20 +28,28 @@ public class FilmService {
 
     private boolean checkValidData(Film film) {
         if (film.getDescription().length() > 200) {
-            log.error("Описание больше 200 символов.", InMemoryFilmStorage.class);
+            log.error("Описание больше 200 символов.", FilmService.class);
             throw new ValidationException("Описание не должно составлять больше 200 символов.");
         }
         if (film.getDescription().isBlank() || film.getDescription().isEmpty()) {
             log.error("Описание пустое или состоит из пробелов.");
-            throw new ValidationException("Описание не должно быть пустым или состоять из пробелов.");
+            throw new ValidationException("Описание не должно быть пустым или состоять из " +
+                    "пробелов.");
         }
         if (film.getReleaseDate().isBefore(MOVIE_BIRTHDAY)) {
-            log.error("Дата выхода фильма в прокат не может быть раньше 28.12.1895.", InMemoryFilmStorage.class);
-            throw new ValidationException("Дата выхода фильма в прокат не может быть раньше 28.12.1895.");
+            log.error("Дата выхода фильма в прокат не может быть раньше 28.12.1895.",
+                    FilmService.class);
+            throw new ValidationException("Дата выхода фильма в прокат не может быть раньше " +
+                    "28.12.1895.");
         }
         if (film.getDuration() <= 0) {
-            log.error("Продолжительность фильма отрицательное число или равно нулю.", InMemoryFilmStorage.class);
+            log.error("Продолжительность фильма отрицательное число или равно нулю.",
+                    FilmService.class);
             throw new ValidationException("Продолжительность фильма должна быть больше нуля.");
+        }
+        if (film.getMpa() == null) {
+            log.error("Не задан рейтинг фильма.");
+            throw new ValidationException("Должен быть задан рейтинг фильма.");
         }
         return true;
     }
@@ -49,7 +57,7 @@ public class FilmService {
     private boolean checkAddValidData(Film film) {
         if (checkValidData(film)) {
             if (film.getName().isBlank()) {
-                log.error("Название пустое.", InMemoryFilmStorage.class);
+                log.error("Название пустое.", FilmService.class);
                 throw new ValidationException("Название не может быть пустым.");
             }
         }
@@ -59,11 +67,11 @@ public class FilmService {
     private boolean checkUpdateValidData(Film film) {
         if (checkValidData(film)) {
             if (film.getId() == null) {
-                log.error("Не введен id.", InMemoryFilmStorage.class);
+                log.error("Не введен id.", FilmService.class);
                 throw new ValidationException("Нужно задать id.");
             }
             if (!getAllFilms().containsKey(film.getId())) {
-                log.error("Неверный id.", InMemoryFilmStorage.class);
+                log.error("Неверный id.", FilmService.class);
                 throw new ElementNotFoundException("фильм с id" + film.getId());
             }
         }
@@ -122,5 +130,9 @@ public class FilmService {
             throw new ElementNotFoundException("фильм " + id);
         }
         return filmStorage.getFilm(id);
+    }
+
+    public Collection<Film> getRecommendations(Integer userId, Integer from, Integer size) {
+        return filmStorage.getRecommendations(userId, from, size);
     }
 }
