+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -21,4 +21,6 @@ public interface FilmStorage {
     Integer addLike(Integer filmId, Integer userId);
 
     Integer removeLike(Integer filmId, Integer userId);
+
+    Collection<Film> getRecommendations(Integer userId, Integer from, Integer size);
 }
