+++ src/test/java/ru/yandex/practicum/filmorate/controllers/UserControllerTest.java
@@ -3,10 +3,11 @@ package ru.yandex.practicum.filmorate.controllers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.jdbc.core.JdbcTemplate;
-import ru.yandex.practicum.filmorate.dao.FriendDao;
-import ru.yandex.practicum.filmorate.dao.UserDbStorage;
+import ru.yandex.practicum.filmorate.dao.FilmDbStorage;
+import ru.yandex.practicum.filmorate.dao.impl.LikesDaoImpl;
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.service.FilmService;
 import ru.yandex.practicum.filmorate.service.UserService;
 import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
@@ -31,7 +32,9 @@ class UserControllerTest {
     @BeforeEach
     public void start() {
         UserStorage userStorage = new InMemoryUserStorage();
-        controller = new UserController(new UserService(userStorage));
+        FilmService filmService = new FilmService(new FilmDbStorage(new JdbcTemplate(),
+                new LikesDaoImpl(new JdbcTemplate())));
+        controller = new UserController(new UserService(userStorage, filmService));
         createUsersForTests();
     }
