+++ src/test/java/ru/yandex/practicum/filmorate/dao/FilmDbStorageTest.java
@@ -12,7 +12,6 @@ import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.Mpa;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -25,6 +24,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
 class FilmDbStorageTest {
     private final FilmDbStorage filmDbStorage;
+
+    private final LikesDao likesDao;
     private Film film;
     private Film film1;
 
@@ -98,7 +99,7 @@ class FilmDbStorageTest {
     @Test
     void shouldUpdateFilmWhenFilmHasNotGenres() {
         Film updateFilm = new Film(1, "updateName", "updateDescription",
-                LocalDate.of(2000, 12, 12),100, new Mpa(4),
+                LocalDate.of(2000, 12, 12), 100, new Mpa(4),
                 new HashSet<>(), null);
         filmDbStorage.updateFilm(updateFilm);
         assertThat(updateFilm).isEqualTo(filmDbStorage.getAllFilms().get(1));
@@ -108,7 +109,7 @@ class FilmDbStorageTest {
     void shouldUpdateFilmWhenFilmHasGenres() {
         filmDbStorage.addFilm(film1);
         Film updateFilm = new Film(2, "updateName", "updateDescription",
-                LocalDate.of(2000, 12, 12),100, new Mpa(4),
+                LocalDate.of(2000, 12, 12), 100, new Mpa(4),
                 new HashSet<>(), new HashSet<>(List.of(new Genre(3), new Genre(4))));
         filmDbStorage.updateFilm(updateFilm);
         assertThat(updateFilm).isEqualTo(filmDbStorage.getAllFilms().get(2));
@@ -126,7 +127,21 @@ class FilmDbStorageTest {
     void shouldReturnFilmById() {
         assertThat(filmDbStorage.getFilm(1)).isEqualTo(new Film(1, "The Rock",
                 "Starring Nicolas Cage and Sean Connery",
-                LocalDate.of(1996, 6, 7),136, new Mpa(1),
+                LocalDate.of(1996, 6, 7), 136, new Mpa(1),
                 new HashSet<>(), null));
+    }
+
+    @Test
+    void shouldReturnPopularFilms() {
+        Film film = new Film(2, "The Rock1", "Starring Nicolas Cage and Sean" +
+                " Connery1", LocalDate.of(1995, 6, 7), 137, new Mpa(2),
+                new HashSet<>());
+        filmDbStorage.addFilm(film);
+        likesDao.addLike(2, 1);
+        assertThat(filmDbStorage.getPopularFilms(2)).isEqualTo(List.of(new Film(2, "The Rock1",
+                "Starring Nicolas Cage and Sean Connery1", LocalDate.of(1995, 6, 7),
+                137, new Mpa(2), new HashSet<>(List.of(1)), null), new Film(1, "The Rock",
+                "Starring Nicolas Cage and Sean Connery", LocalDate.of(1996, 6, 7),
+                136, new Mpa(1), null)));
     }
 }
