+++ src/test/java/ru/yandex/practicum/filmorate/dao/LikesDaoTest.java
@@ -6,13 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.model.Mpa;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -38,18 +31,5 @@ class LikesDaoTest {
         assertThat(filmDbStorage.getAllFilms().get(1).getLikes()).contains(1);
         likesDao.removeLike(1, 1);
         assertThat(filmDbStorage.getAllFilms().get(1).getLikes()).isEmpty();
-    }
-
-    @Test
-    void shouldReturnPopularFilms() {
-        Film film = new Film(2,"The Rock1", "Starring Nicolas Cage and Sean Connery1",
-                LocalDate.of(1995, 6, 7), 137, new Mpa(2), new HashSet<>());
-        filmDbStorage.addFilm(film);
-        likesDao.addLike(2, 1);
-        assertThat(likesDao.getPopularFilms(2)).isEqualTo(List.of(new Film(2,"The Rock1",
-                "Starring Nicolas Cage and Sean Connery1", LocalDate.of(1995, 6, 7),
-                137, new Mpa(2), new HashSet<>(List.of(1)), new HashSet<>()), new Film(1,"The Rock",
-                "Starring Nicolas Cage and Sean Connery", LocalDate.of(1996, 6, 7),
-                136, new Mpa(1), new HashSet<>())));
     }
 }
